Finalizado escopo de apostas inicial

+++ src/Monkify.Domain/Sessions/Entities/SessionParameters.cs
@@ -11,7 +11,7 @@ namespace Monkify.Domain.Sessions.Entities
     public class SessionParameters : TableEntity
     {
         public SessionCharacterType SessionCharacterType { get; set; }
-        public double RequiredAmount { get; set; }
+        public decimal RequiredAmount { get; set; }
         public int MinimumNumberOfPlayers { get; set; }
         public int ChoiceRequiredLength { get; set; }
         public bool AcceptDuplicatedCharacters { get; set; }
